@@ -10,10 +10,10 @@
     }
   ],
   "require": {
-    "php": ">=7.2",
+    "php": ">=7.4",
     "sauls/collections": "^1.0",
     "sauls/options-resolver": "^1.0",
-    "twig/twig": "^2.0 || ^3.0"
+    "twig/twig": "^3.0"
   },
   "require-dev": {
     "phpunit/phpunit": "^7.0",
